challenges/Blind Hacker @ Fwhibbit CTR 2020/blind-hacker/flask/requirements.txt
Flask
pyjwt
shellescape
pymysql
psycopg2
cryptography
